rukito-backend/analytics/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
requests
beautifulsoup4
python-dotenv
cryptography